integrations/mcp/python/kukapay_info_server.py
"""
Kukapay Info MCP Server - Read-only Hyperliquid data access
Port: 8401
Real data from Hyperliquid API
"""

import asyncio
import json
import os
import sys
from datetime import datetime, timedelta
from typing import Dict, Any, Optional, List
from pathlib import Path

# Add parent directory to path
sys.path.append(str(Path(__file__).parent.parent.parent))

from fastapi import FastAPI, HTTPException
from pydantic import BaseModel, Field
import uvicorn
import requests

# Try to import Hyperliquid SDK
try:
    from hyperliquid.info import Info
    from hyperliquid.utils import constants
    HAS_SDK = True
except ImportError:
    HAS_SDK = False
    print("Warning: Hyperliquid SDK not found, using direct API calls")


class MCPRequest(BaseModel):
    method: str
    params: Dict[str, Any] = Field(default_factory=dict)
    id: Optional[str] = None


class MCPResponse(BaseModel):
    result: Optional[Any] = None
    error: Optional[Dict[str, Any]] = None
    id: Optional[str] = None


class KukapayInfoServer:
    """Kukapay Info MCP Server - Analytics focused"""
    
    def __init__(self, port: int = 8401):
        self.app = FastAPI(title="Kukapay Info MCP Server", version="1.0.0")
        self.port = port
        self.setup_routes()
        
        if HAS_SDK:
            self.info = Info(constants.MAINNET_API_URL, skip_ws=True)
        else:
            self.info = None
        
        self.api_url = "https://api.hyperliquid.xyz/info"
        
    def setup_routes(self):
        @self.app.get("/")
        async def root():
            return {
                "name": "Kukapay Info MCP Server",
                "version": "1.0.0",
                "port": self.port,
                "data_source": "REAL - Hyperliquid Mainnet",
                "features": ["analytics", "read-only", "market-data"],
                "ready": True
            }
        
        @self.app.get("/health")
        async def health():
            return {"status": "healthy", "timestamp": datetime.now().isoformat()}
        
        @self.app.get("/mcp/tools")
        async def list_tools():
            return {"tools": self.get_available_tools()}
        
        @self.app.post("/mcp/execute")
        async def execute_tool(request: MCPRequest):
            try:
                result = await self.execute_method(request.method, request.params)
                return MCPResponse(result=result, id=request.id)
            except Exception as e:
                return MCPResponse(error={"code": -1, "message": str(e)}, id=request.id)
    
    def get_available_tools(self) -> List[Dict[str, Any]]:
        return [
            {
                "name": "get_user_state",
                "description": "Get user account state and balances",
                "parameters": {"address": {"type": "string", "required": True}}
            },
            {
                "name": "get_user_open_orders",
                "description": "Get user's open orders",
                "parameters": {"address": {"type": "string", "required": True}}
            },
            {
                "name": "analyze_positions",
                "description": "Analyze user positions with PnL and risk metrics",
                "parameters": {"address": {"type": "string", "required": True}}
            },
            {
                "name": "get_leaderboard",
                "description": "Get trading leaderboard",
                "parameters": {"timeframe": {"type": "string", "default": "day"}}
            },
            {
                "name": "get_liquidations",
                "description": "Get recent liquidations",
                "parameters": {"lookback_hours": {"type": "number", "default": 24}}
            },
            {
                "name": "get_market_summary",
                "description": "Get comprehensive market summary",
                "parameters": {}
            }
        ]
    
    async def execute_method(self, method: str, params: Dict[str, Any]) -> Any:
        method_map = {
            "get_user_state": self.get_user_state,
            "get_user_open_orders": self.get_user_open_orders,
            "analyze_positions": self.analyze_positions,
            "get_leaderboard": self.get_leaderboard,
            "get_liquidations": self.get_liquidations,
            "get_market_summary": self.get_market_summary
        }
        
        if method not in method_map:
            raise ValueError(f"Unknown method: {method}")
        
        return await method_map[method](params)
    
    async def get_user_state(self, params: Dict[str, Any]) -> Dict[str, Any]:
        """Get real user state from Hyperliquid"""
        address = params.get("address")
        if not address:
            return {"error": "Address required"}
        
        try:
            if self.info:
                result = self.info.user_state(address)
                return result if result else {"address": address, "error": "No data found"}
            else:
                response = requests.post(
                    self.api_url,
                    json={"type": "clearinghouseState", "user": address},
                    timeout=10
                )
                if response.status_code == 200:
                    return response.json()
                return {"address": address, "error": "API error"}
        except Exception as e:
            return {"address": address, "error": str(e)}
    
    async def get_user_open_orders(self, params: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Get real open orders for a user"""
        address = params.get("address")
        if not address:
            return []
        
        try:
            if self.info:
                result = self.info.open_orders(address)
                return result if result else []
            else:
                response = requests.post(
                    self.api_url,
                    json={"type": "openOrders", "user": address},
                    timeout=10
                )
                if response.status_code == 200:
                    return response.json()
                return []
        except Exception as e:
            print(f"Error fetching open orders: {e}")
            return []
    
    async def analyze_positions(self, params: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze user positions with detailed metrics"""
        address = params.get("address")
        if not address:
            return {"error": "Address required"}
        
        try:
            # Get user state
            user_state = await self.get_user_state(params)
            
            # Get current prices
            response = requests.post(
                self.api_url,
                json={"type": "allMids"},
                timeout=10
            )
            prices = response.json() if response.status_code == 200 else {}
            
            # Analyze positions
            positions = []
            total_pnl = 0
            total_exposure = 0
            
            if "assetPositions" in user_state:
                for pos in user_state.get("assetPositions", []):
                    if isinstance(pos, dict):
                        position = pos.get("position", {})
                        coin = position.get("coin", "")
                        size = float(position.get("szi", 0))
                        entry_px = float(position.get("entryPx", 0))
                        current_px = float(prices.get(coin, entry_px))
                        
                        if size != 0:
                            pnl = (current_px - entry_px) * size
                            total_pnl += pnl
                            total_exposure += abs(size * current_px)
                            
                            positions.append({
                                "coin": coin,
                                "size": size,
                                "entry_price": entry_px,
                                "current_price": current_px,
                                "pnl": pnl,
                                "pnl_percent": ((current_px / entry_px - 1) * 100) if entry_px > 0 else 0
                            })
            
            return {
                "address": address,
                "positions": positions,
                "total_pnl": total_pnl,
                "total_exposure": total_exposure,
                "position_count": len(positions),
                "timestamp": datetime.now().isoformat()
            }
        except Exception as e:
            return {"address": address, "error": str(e)}
    
    async def get_leaderboard(self, params: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Get trading leaderboard (simulated for now)"""
        timeframe = params.get("timeframe", "day")
        
        # Note: Real leaderboard would require specific API endpoint
        # This is a placeholder showing the structure
        return [
            {
                "rank": 1,
                "address": "0x1234...5678",
                "pnl": 125000.50,
                "roi": 25.5,
                "trades": 150,
                "timeframe": timeframe
            },
            {
                "rank": 2,
                "address": "0xabcd...efgh",
                "pnl": 98000.25,
                "roi": 18.2,
                "trades": 89,
                "timeframe": timeframe
            }
        ]
    
    async def get_liquidations(self, params: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Get recent liquidations (structure example)"""
        lookback_hours = params.get("lookback_hours", 24)
        
        # Note: Real liquidations would require specific API endpoint
        return {
            "lookback_hours": lookback_hours,
            "total_liquidations": 0,
            "total_volume": 0,
            "liquidations": [],
            "note": "Liquidation endpoint requires additional API access"
        }
    
    async def get_market_summary(self, params: Dict[str, Any]) -> Dict[str, Any]:
        """Get comprehensive market summary"""
        try:
            # Get all prices
            response = requests.post(
                self.api_url,
                json={"type": "allMids"},
                timeout=10
            )
            prices = response.json() if response.status_code == 200 else {}
            
            # Calculate market metrics
            top_gainers = []
            top_losers = []
            total_coins = len(prices)
            
            # Get metadata for market cap info
            meta_response = requests.post(
                self.api_url,
                json={"type": "meta"},
                timeout=10
            )
            meta = meta_response.json() if meta_response.status_code == 200 else {}
            
            return {
                "total_coins": total_coins,
                "top_prices": {k: v for k, v in list(prices.items())[:10]},
                "market_status": "open",
                "last_update": datetime.now().isoformat(),
                "data_source": "real"
            }
        except Exception as e:
            return {"error": str(e)}


def run_server(port: int = 8401):
    server = KukapayInfoServer(port)
    print(f"Starting Kukapay Info MCP Server on port {port}")
    print(f"Data Source: REAL Hyperliquid Mainnet")
    print(f"Health: http://localhost:{port}/health")
    uvicorn.run(server.app, host="0.0.0.0", port=port)


if __name__ == "__main__":
    run_server(8401)